add fixes from review
