+++ assume/common/utils.py
@@ -15,6 +15,7 @@ import dateutil.rrule as rr
 import numpy as np
 import pandas as pd
 
+from assume.common.base import BaseStrategy, LearningStrategy
 from assume.common.market_objects import MarketProduct, Orderbook
 
 logger = logging.getLogger(__name__)
@@ -471,3 +472,36 @@ def convert_to_rrule_freq(string: str) -> tuple[int, int]:
     freq = freq_map[string[-1]]
     interval = int(string[:-1])
     return freq, interval
+
+
+def adjust_unit_operator_for_learning(
+    bidding_strategies: dict[str, str],
+    world_bidding_strategies: dict[str, BaseStrategy],
+    unit_operator_id: str,
+):
+    """
+    Check if any of the bidding strategies are learning strategies.
+    And change the unit operator to RL if learning strategies are found.
+
+    Args:
+        bidding_strategies (dict[str, str]): The bidding strategies for the unit.
+        world_bidding_strategies (dict[str, BaseStrategy]): The bidding strategies of the World
+        unit_operator_id (str): The identifier of the unit operator.
+
+    Returns:
+        str: The corrected unit operator identifier.
+
+    """
+    if unit_operator_id == "Operator-RL":
+        return unit_operator_id
+    for strategy in bidding_strategies.values():
+        if issubclass(world_bidding_strategies[strategy], LearningStrategy):
+            unit_operator_id = "Operator-RL"
+            logger.debug(
+                "Your chosen unit operator %s for the learning unit %s was overwritten with 'Operator-RL', "
+                "since all learning units need to be handeled by one unit operator.",
+                unit_operator_id,
+                id,
+            )
+
+    return unit_operator_id
